requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Pillow==10.4.0
gunicorn==21.2.0
requests==2.31.0
